% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{body_add_from_spec}
\alias{body_add_from_spec}
\title{Create a list of parallel tasks from a specification}
\usage{
body_add_from_spec(doc, spec_list, task_adder = NULL, ...)
}
\arguments{
\item{doc}{An officer document object to be passed to each future function.}

\item{spec_list}{A list of lists, where each inner list specifies a task. Each inner list must have a 'type' property defining which future function to call (e.g. 'caption', 'flextable', 'text', 'plot', 'element_list').
It may also include a 'value' property, which is a function that takes a data argument and ... arguments. The value function is wrapped so that the ... passed to body_add_from_spec are forwarded to it.
Any additional properties will be passed as extra arguments to the future function.}

\item{...}{Additional arguments that will be passed to the value functions of tasks requiring a 'value' argument.}
}
\value{
A list of task objects (each a list with 'task' and 'callback') that can be executed using parallel_task_queue.
}
\description{
Create a list of parallel tasks from a specification
}
