% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{configure_logging}
\alias{configure_logging}
\title{Configure logging settings}
\usage{
configure_logging(level = LOG_LEVELS$INFO, file = NULL)
}
\arguments{
\item{level}{Logging level (one of LOG_LEVELS)}

\item{file}{Optional file path to write logs to}
}
\description{
Configure logging settings
}
