% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{future_body_add_caption}
\alias{future_body_add_caption}
\title{Create a parallel task for adding a caption to an officer document}
\usage{
future_body_add_caption(x, value, ...)
}
\arguments{
\item{x}{An officer document object.}

\item{value}{A caption text or an expression resulting in caption text.}

\item{...}{Additional arguments passed to officer::body_add_caption}
}
\value{
A list with elements 'task' and 'callback'.
}
\description{
This function has the same interface as officer::body_add_caption. It returns a task that can be executed in parallel using parallel_task_queue.
The task evaluates the caption text and returns it, and the callback adds the caption to the document.
}
