% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{future_body_add_flextable}
\alias{future_body_add_flextable}
\title{Create a parallel task for adding a flextable to an officer document}
\usage{
future_body_add_flextable(x, value, ...)
}
\arguments{
\item{x}{An officer document object.}

\item{value}{A flextable object or an expression resulting in a flextable.}

\item{...}{Additional arguments passed to flextable::body_add_flextable.}
}
\value{
A list with elements 'task' and 'callback'.
}
\description{
This function has the same interface as flextable::body_add_flextable. It returns a task that can be executed in parallel using parallel_task_queue.
The task evaluates the flextable object and returns it, and the callback adds the flextable to the document.
}
