% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{future_body_add_fpar}
\alias{future_body_add_fpar}
\title{Create a parallel task for adding a formatted paragraph (fpar) to an officer document}
\usage{
future_body_add_fpar(x, value, ...)
}
\arguments{
\item{x}{An officer document object.}

\item{value}{An fpar object or an expression resulting in an fpar.}

\item{...}{Additional arguments passed to officer::body_add_fpar.}
}
\value{
A list with elements 'task' and 'callback'.
}
\description{
This function has the same interface as officer::body_add_fpar. It returns a task that can be executed in parallel using parallel_task_queue.
The task evaluates the formatted paragraph object and returns it, and the callback adds it to the document.
}
