% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{future_body_add_par}
\alias{future_body_add_par}
\title{Create a parallel task for adding a paragraph to an officer document}
\usage{
future_body_add_par(x, value, ...)
}
\arguments{
\item{x}{An officer document object.}

\item{value}{A character string or expression resulting in text.}

\item{...}{Additional arguments passed to officer::body_add_par.}
}
\value{
A list with elements 'task' and 'callback'.
}
\description{
This function has the same interface as officer::body_add_par. It returns a task that can be executed in parallel using parallel_task_queue.
The task evaluates the paragraph text and returns it, and the callback adds the paragraph to the document.
}
