% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.R
\name{future_body_end_section_landscape}
\alias{future_body_end_section_landscape}
\title{Create a parallel task for ending a section with landscape orientation in an officer document}
\usage{
future_body_end_section_landscape(x, ...)
}
\arguments{
\item{x}{An officer document object.}

\item{...}{Additional arguments passed to officer::body_end_section_landscape.}
}
\value{
A list with elements 'task' and 'callback'.
}
\description{
This function has the same interface as officer::body_end_section_landscape. It returns a task that can be executed in parallel using parallel_task_queue.
The task does not perform a computation, and the callback ends the section with landscape orientation.
}
