% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_message}
\alias{log_message}
\title{Internal logging function}
\usage{
log_message(level, msg, ...)
}
\arguments{
\item{level}{The log level for this message}

\item{msg}{The message to log}

\item{...}{Additional objects to include in log message}
}
\description{
Internal logging function
}
