% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\docType{data}
\name{logging_config}
\alias{logging_config}
\title{Global logging configuration}
\format{
An object of class \code{environment} of length 2.
}
\usage{
logging_config
}
\description{
Global logging configuration
}
\keyword{datasets}
